Refactored, adjusted code using commentary from v2 pull
