+++ dino_catalog/print_test.rb
@@ -5,7 +5,7 @@ dino_files = FileFinder.new(".", "*.csv")
 dino_files.files.each { |file_path| dino_dex.import_csv_file(file_path) }
 
 puts "[Bipeds]"
-dino_dex.select("walk_type", "biped").print_data("name")
+dino_dex.select("walking", "biped").print_data("name")
 puts "`````````````"
 
 puts "[Carnivores]"
@@ -29,7 +29,7 @@ dino_dex.select_smaller_than(500).print_data(["name", "weight"])
 puts "`````````````"
 
 puts "[Dinosaur Facts -> Yangchuanosaurus]"
-dino_dex.select_dinosaur("yangchuanosaurus").print_facts unless dino_dex.select_dinosaur("yangchuanosaurus").nil?
+puts dino_dex.select_dinosaur("yangchuanosaurus").export_facts
 puts "`````````````"
 
-puts dino_dex.export_json
+puts dino_dex.to_json
